Move `base_comparator` and `load_comparator` to the `wft4galaxy.comparators` package
